Regression/requirements.txt
mlflow>=2.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
databricks-sdk>=0.20.0
cloudpickle>=2.2.0
